Setup all tests and skip failing ones
